+++ planetmint/__init__.py
@@ -8,7 +8,7 @@ from transactions.types.elections.validator_election import ValidatorElection  #
 from transactions.types.elections.vote import Vote  # noqa
 from transactions.types.elections.chain_migration_election import ChainMigrationElection
 from planetmint.lib import Planetmint
-from planetmint.core import App
+from planetmint.abci.core import App
 
 
 Transaction.register_type(Transaction.CREATE, Transaction)
