+++ planetmint/abci/parallel_validation.py
@@ -8,7 +8,7 @@ import multiprocessing
 from collections import defaultdict
 from planetmint import App
 from planetmint.lib import Planetmint
-from planetmint.tendermint_utils import decode_transaction
+from planetmint.abci.tendermint_utils import decode_transaction
 from abci.application import OkCode
 from tendermint.abci.types_pb2 import (
     ResponseCheckTx,
