+++ planetmint/commands/planetmint.py
@@ -14,7 +14,7 @@ import json
 import sys
 import planetmint
 
-from planetmint.core import rollback
+from planetmint.abci.core import rollback
 from planetmint.utils import load_node_key
 from transactions.common.transaction_mode_types import BROADCAST_TX_COMMIT
 from transactions.common.exceptions import DatabaseDoesNotExist, ValidationError
@@ -27,7 +27,7 @@ from planetmint.backend import schema
 from planetmint.commands import utils
 from planetmint.commands.utils import configure_planetmint, input_on_stderr
 from planetmint.log import setup_logging
-from planetmint.tendermint_utils import public_key_from_base64
+from planetmint.abci.tendermint_utils import public_key_from_base64
 from planetmint.commands.election_types import elections
 from planetmint.version import __tm_supported_versions__
 from planetmint.config import Config
