+++ tests/tendermint/test_core.py
@@ -17,12 +17,11 @@ from transactions.types.assets.create import Create
 from transactions.types.assets.transfer import Transfer
 from planetmint import App
 from planetmint.backend import query
-from planetmint.core import OkCode, CodeTypeError, rollback
+from planetmint.abci.core import OkCode, CodeTypeError, rollback
 from planetmint.lib import Block
-from planetmint.tendermint_utils import new_validator_set
-from planetmint.tendermint_utils import public_key_to_base64
+from planetmint.abci.tendermint_utils import new_validator_set
+from planetmint.abci.tendermint_utils import public_key_to_base64
 from planetmint.version import __tm_supported_versions__
-from planetmint.backend.tarantool.const import TARANT_TABLE_GOVERNANCE
 from tests.utils import generate_election, generate_validators
 
 pytestmark = pytest.mark.bdb
