+++ tests/tendermint/test_integration.py
@@ -20,7 +20,7 @@ from io import BytesIO
 @pytest.mark.bdb
 def test_app(b, eventqueue_fixture, init_chain_request):
     from planetmint import App
-    from planetmint.tendermint_utils import calculate_hash
+    from planetmint.abci.tendermint_utils import calculate_hash
     from transactions.common.crypto import generate_key_pair
 
     app = App(b, eventqueue_fixture)
