+++ tests/upsert_validator/test_upsert_validator_vote.py
@@ -6,7 +6,7 @@
 import pytest
 import codecs
 
-from planetmint.tendermint_utils import public_key_to_base64
+from planetmint.abci.tendermint_utils import public_key_to_base64
 
 from transactions.types.elections.validator_election import ValidatorElection
 from transactions.common.exceptions import AmountError
