--- tmdata/config.toml
@@ -1,18 +1,0 @@
-# This is a TOML config file.
-# For more information, see https://github.com/toml-lang/toml
-
-proxy_app = "tcp://planetmint:26658"
-moniker = "anonymous"
-fast_sync = true
-db_backend = "leveldb"
-log_level = "state:debug,*:error"
-
-[consensus]
-create_empty_blocks = false
-
-[rpc]
-laddr = "tcp://0.0.0.0:26657"
-
-[p2p]
-laddr = "tcp://0.0.0.0:26656"
-seeds = ""
